+++ lib/ccru.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
+require_relative "ccru/cli"
 require_relative "ccru/version"
 
 module Ccru
   class Error < StandardError; end
-  # Your code goes here...
 end